+++ internal/background/worker.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"runtime/pprof"
 	"sync"
-	"sync/atomic"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/empty"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xcontext"
@@ -24,12 +23,18 @@ type Worker struct {
 	workers  sync.WaitGroup
 	onceInit sync.Once
 
+	tasksCompleted empty.Chan
+
 	m xsync.Mutex
 
-	closed      uint32
+	tasks chan backgroundTask
+
+	closed      bool
 	stop        xcontext.CancelErrFunc
 	closeReason error
 }
+
+type CallbackFunc func(ctx context.Context)
 
 func NewWorker(parent context.Context) *Worker {
 	w := Worker{}
@@ -44,54 +49,52 @@ func (b *Worker) Context() context.Context {
 	return b.ctx
 }
 
-func (b *Worker) Start(name string, f func(ctx context.Context)) {
-	if atomic.LoadUint32(&b.closed) != 0 {
-		f(b.ctx)
-		return
-	}
-
+func (b *Worker) Start(name string, f CallbackFunc) {
 	b.init()
 
-	b.m.Lock()
-	defer b.m.Unlock()
+	b.m.WithLock(func() {
+		if b.closed {
+			return
+		}
 
-	if b.ctx.Err() != nil {
-		return
-	}
-
-	b.workers.Add(1)
-	go func() {
-		defer b.workers.Done()
-
-		pprof.Do(b.ctx, pprof.Labels("background", name), f)
-	}()
+		b.tasks <- backgroundTask{
+			callback: f,
+			name:     name,
+		}
+	})
 }
 
 func (b *Worker) Done() <-chan struct{} {
 	b.init()
 
-	b.m.Lock()
-	defer b.m.Unlock()
-
 	return b.ctx.Done()
 }
 
 func (b *Worker) Close(ctx context.Context, err error) error {
-	if !atomic.CompareAndSwapUint32(&b.closed, 0, 1) {
-		return xerrors.WithStackTrace(ErrAlreadyClosed)
-	}
-
 	b.init()
 
-	b.m.Lock()
-	defer b.m.Unlock()
+	var resErr error
+	b.m.WithLock(func() {
+		if b.closed {
+			resErr = xerrors.WithStackTrace(ErrAlreadyClosed)
+			return
+		}
 
-	b.closeReason = err
-	if b.closeReason == nil {
-		b.closeReason = errClosedWithNilReason
+		b.closed = true
+
+		close(b.tasks)
+		b.closeReason = err
+		if b.closeReason == nil {
+			b.closeReason = errClosedWithNilReason
+		}
+
+		b.stop(err)
+	})
+	if resErr != nil {
+		return resErr
 	}
 
-	b.stop(err)
+	<-b.tasksCompleted
 
 	bgCompleted := make(empty.Chan)
 
@@ -120,5 +123,27 @@ func (b *Worker) init() {
 		if b.ctx == nil {
 			b.ctx, b.stop = xcontext.WithErrCancel(context.Background())
 		}
+		b.tasks = make(chan backgroundTask)
+		b.tasksCompleted = make(empty.Chan)
+		go b.starterLoop()
 	})
+}
+
+func (b *Worker) starterLoop() {
+	defer close(b.tasksCompleted)
+
+	for bgTask := range b.tasks {
+		b.workers.Add(1)
+
+		go func(task backgroundTask) {
+			defer b.workers.Done()
+
+			pprof.Do(b.ctx, pprof.Labels("background", task.name), task.callback)
+		}(bgTask)
+	}
+}
+
+type backgroundTask struct {
+	callback CallbackFunc
+	name     string
 }