+++ table/options/options_test.go
@@ -23,7 +23,7 @@ func TestSessionOptionsProfile(t *testing.T) {
 			WithProfilePreset(abc),
 		)
 		req := Ydb_Table.CreateTableRequest{}
-		opt((*CreateTableDesc)(&req), a)
+		opt.ApplyCreateTableOption((*CreateTableDesc)(&req), a)
 		if req.Profile.PresetName != abc {
 			t.Errorf("Preset is not as expected")
 		}
@@ -33,7 +33,7 @@ func TestSessionOptionsProfile(t *testing.T) {
 			WithCompactionPolicy(WithCompactionPolicyPreset(abc)),
 		)
 		req := Ydb_Table.CreateTableRequest{}
-		opt((*CreateTableDesc)(&req), a)
+		opt.ApplyCreateTableOption((*CreateTableDesc)(&req), a)
 		if req.Profile.CompactionPolicy.PresetName != abc {
 			t.Errorf("Compaction policy is not as expected")
 		}
@@ -46,7 +46,7 @@ func TestSessionOptionsProfile(t *testing.T) {
 			),
 		)
 		req := Ydb_Table.CreateTableRequest{}
-		opt((*CreateTableDesc)(&req), a)
+		opt.ApplyCreateTableOption((*CreateTableDesc)(&req), a)
 		p := req.Profile.PartitioningPolicy
 		if p.PresetName != abc || p.AutoPartitioning != Ydb_Table.PartitioningPolicy_AUTO_SPLIT {
 			t.Errorf("Partitioning policy is not as expected")
@@ -59,7 +59,7 @@ func TestSessionOptionsProfile(t *testing.T) {
 			),
 		)
 		req := Ydb_Table.CreateTableRequest{}
-		opt((*CreateTableDesc)(&req), a)
+		opt.ApplyCreateTableOption((*CreateTableDesc)(&req), a)
 		p := req.Profile.PartitioningPolicy
 		if pp, ok := p.Partitions.(*Ydb_Table.PartitioningPolicy_UniformPartitions); !ok || pp.UniformPartitions != 3 {
 			t.Errorf("Uniform partitioning policy is not as expected")
@@ -72,7 +72,7 @@ func TestSessionOptionsProfile(t *testing.T) {
 			),
 		)
 		req := Ydb_Table.CreateTableRequest{}
-		opt((*CreateTableDesc)(&req), a)
+		opt.ApplyCreateTableOption((*CreateTableDesc)(&req), a)
 		p := req.Profile.PartitioningPolicy
 
 		pp, ok := p.Partitions.(*Ydb_Table.PartitioningPolicy_ExplicitPartitions)
@@ -91,7 +91,7 @@ func TestSessionOptionsProfile(t *testing.T) {
 			WithExecutionPolicy(WithExecutionPolicyPreset(abc)),
 		)
 		req := Ydb_Table.CreateTableRequest{}
-		opt((*CreateTableDesc)(&req), a)
+		opt.ApplyCreateTableOption((*CreateTableDesc)(&req), a)
 		if req.Profile.ExecutionPolicy.PresetName != abc {
 			t.Errorf("Execution policy is not as expected")
 		}
@@ -106,7 +106,7 @@ func TestSessionOptionsProfile(t *testing.T) {
 			),
 		)
 		req := Ydb_Table.CreateTableRequest{}
-		opt((*CreateTableDesc)(&req), a)
+		opt.ApplyCreateTableOption((*CreateTableDesc)(&req), a)
 		p := req.Profile.ReplicationPolicy
 		if p.PresetName != abc ||
 			p.ReplicasCount != 3 ||
@@ -120,7 +120,7 @@ func TestSessionOptionsProfile(t *testing.T) {
 			WithCachingPolicy(WithCachingPolicyPreset(abc)),
 		)
 		req := Ydb_Table.CreateTableRequest{}
-		opt((*CreateTableDesc)(&req), a)
+		opt.ApplyCreateTableOption((*CreateTableDesc)(&req), a)
 		if req.Profile.CachingPolicy.PresetName != abc {
 			t.Errorf("Caching policy is not as expected")
 		}
@@ -142,7 +142,7 @@ func TestStoragePolicyOptions(t *testing.T) {
 			),
 		)
 		req := Ydb_Table.CreateTableRequest{}
-		opt((*CreateTableDesc)(&req), a)
+		opt.ApplyCreateTableOption((*CreateTableDesc)(&req), a)
 		p := req.Profile.StoragePolicy
 		if p.PresetName != abc ||
 			p.Syslog.Media != "any1" ||
@@ -161,7 +161,7 @@ func TestAlterTableOptions(t *testing.T) {
 	{
 		opt := WithAddColumn("a", types.TypeBool)
 		req := Ydb_Table.AlterTableRequest{}
-		opt((*AlterTableDesc)(&req), a)
+		opt.ApplyAlterTableOption((*AlterTableDesc)(&req), a)
 		if len(req.AddColumns) != 1 ||
 			req.AddColumns[0].Name != "a" {
 			t.Errorf("Alter table options is not as expected")
@@ -175,7 +175,7 @@ func TestAlterTableOptions(t *testing.T) {
 		}
 		opt := WithAddColumnMeta(column)
 		req := Ydb_Table.AlterTableRequest{}
-		opt((*AlterTableDesc)(&req), a)
+		opt.ApplyAlterTableOption((*AlterTableDesc)(&req), a)
 		if len(req.AddColumns) != 1 ||
 			req.AddColumns[0].Name != column.Name ||
 			req.AddColumns[0].Type != value.TypeToYDB(column.Type, a) ||
@@ -186,7 +186,7 @@ func TestAlterTableOptions(t *testing.T) {
 	{
 		opt := WithDropColumn("a")
 		req := Ydb_Table.AlterTableRequest{}
-		opt((*AlterTableDesc)(&req), a)
+		opt.ApplyAlterTableOption((*AlterTableDesc)(&req), a)
 		if len(req.DropColumns) != 1 ||
 			req.DropColumns[0] != "a" {
 			t.Errorf("Alter table options is not as expected")
@@ -203,7 +203,7 @@ func TestAlterTableOptions(t *testing.T) {
 		}
 		opt := WithAlterColumnFamilies(cf)
 		req := Ydb_Table.AlterTableRequest{}
-		opt((*AlterTableDesc)(&req), a)
+		opt.ApplyAlterTableOption((*AlterTableDesc)(&req), a)
 		if len(req.AddColumnFamilies) != 1 ||
 			req.AddColumnFamilies[0].Name != cf.Name ||
 			req.AddColumnFamilies[0].Data.Media != cf.Data.Media ||
@@ -219,7 +219,7 @@ func TestAlterTableOptions(t *testing.T) {
 		}
 		opt := WithAlterColumnFamilies(cf)
 		req := Ydb_Table.AlterTableRequest{}
-		opt((*AlterTableDesc)(&req), a)
+		opt.ApplyAlterTableOption((*AlterTableDesc)(&req), a)
 		if len(req.AddColumnFamilies) != 1 ||
 			req.AddColumnFamilies[0].Name != cf.Name ||
 			req.AddColumnFamilies[0].Data != nil ||
@@ -235,7 +235,7 @@ func TestAlterTableOptions(t *testing.T) {
 		}
 		opt := WithAlterReadReplicasSettings(rr)
 		req := Ydb_Table.AlterTableRequest{}
-		opt((*AlterTableDesc)(&req), a)
+		opt.ApplyAlterTableOption((*AlterTableDesc)(&req), a)
 		rrOut := NewReadReplicasSettings(req.GetSetReadReplicasSettings())
 		if rr != rrOut {
 			t.Errorf("Alter table set read replicas options is not as expected")
@@ -250,7 +250,7 @@ func TestAlterTableOptions(t *testing.T) {
 		}
 		opt := WithAlterStorageSettings(ss)
 		req := Ydb_Table.AlterTableRequest{}
-		opt((*AlterTableDesc)(&req), a)
+		opt.ApplyAlterTableOption((*AlterTableDesc)(&req), a)
 		rrOut := NewStorageSettings(req.GetAlterStorageSettings())
 		if ss != rrOut {
 			t.Errorf("Alter table storage settings options is not as expected")